feat(ehr): rework
Ref: ENG-139

Ref: #1040
